refactor(initializeHandlers): remove middleware and move logic into initializeHandlers
- remove twitchHandlersMiddleware file
- move initializeAuthToken  into initializeHandlers file
- create init function - replacing middleware
